+++ launcher/pkg/src/main/java/com/blankj/launcher/pkg/SplashActivity.kt
@@ -5,7 +5,6 @@ import android.view.View
 import com.blankj.base.BaseActivity
 
 class SplashActivity : BaseActivity() {
-
     override fun initData(bundle: Bundle?) {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
@@ -14,7 +13,7 @@ class SplashActivity : BaseActivity() {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun initView(savedInstanceState: Bundle?, contentView: View?) {
+    override fun initView(savedInstanceState: Bundle?, contentView: View) {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
@@ -22,7 +21,7 @@ class SplashActivity : BaseActivity() {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun onWidgetClick(view: View?) {
+    override fun onWidgetClick(view: View) {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }