+++ lib/base/src/main/java/com/blankj/base/BaseKotlinActivity.kt
@@ -1,8 +1,0 @@
-package com.blankj.base
-
-import android.support.v7.app.AppCompatActivity
-
-class BaseKotlinActivity : AppCompatActivity() {
-
-
-}