+++ utilcode/lib/src/main/java/com/blankj/utilcode/util/ToastUtils.java
@@ -4,6 +4,7 @@ import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
 import android.content.res.Configuration;
+import android.content.res.Resources;
 import android.graphics.Color;
 import android.graphics.PixelFormat;
 import android.graphics.PorterDuff;
@@ -17,9 +18,9 @@ import android.os.Message;
 import android.support.annotation.ColorInt;
 import android.support.annotation.DrawableRes;
 import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
 import android.support.annotation.StringRes;
 import android.support.v4.app.NotificationManagerCompat;
+import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -250,7 +251,7 @@ public final class ToastUtils {
             @Override
             public void run() {
                 cancel();
-                sToast = ToastFactory.makeToast(Utils.getApp(), text, duration);
+                sToast = ToastFactory.makeToast(Utils.getTopActivityOrApp(), text, duration);
                 final TextView tvMessage = sToast.getView().findViewById(android.R.id.message);
                 if (sMsgColor != COLOR_DEFAULT) {
                     tvMessage.setTextColor(sMsgColor);
@@ -340,21 +341,14 @@ public final class ToastUtils {
             if (NotificationManagerCompat.from(context).areNotificationsEnabled()) {
                 return new SystemToast(makeNormalToast(context, text, duration));
             }
-
             return new ToastWithoutNotification(makeNormalToast(context, text, duration));
-//            Log.e("ToastUtils", "Toast is GG. In fact, next step is useless.");
-//            return new SystemToast(makeNormalToast(context, text, duration));
         }
 
         static IToast newToast(Context context) {
             if (NotificationManagerCompat.from(context).areNotificationsEnabled()) {
                 return new SystemToast(new Toast(context));
             }
-//            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N_MR1) {
             return new ToastWithoutNotification(new Toast(context));
-//            }
-//            Log.e("ToastUtils", "Toast is GG. In fact, next step is useless.");
-//            return new SystemToast(new Toast(context));
         }
 
         private static Toast makeNormalToast(Context context, CharSequence text, int duration) {
@@ -365,15 +359,13 @@ public final class ToastUtils {
         }
     }
 
-    static class SystemToast implements IToast {
-
-        Toast mToast;
+    static class SystemToast extends AbsToast {
 
         private static Field sField_mTN;
         private static Field sField_TN_Handler;
 
-        SystemToast(@NonNull Toast toast) {
-            mToast = toast;
+        SystemToast(Toast toast) {
+            super(toast);
             if (Build.VERSION.SDK_INT == Build.VERSION_CODES.N_MR1) {
                 try {
                     //noinspection JavaReflectionMemberAccess
@@ -398,36 +390,6 @@ public final class ToastUtils {
             mToast.cancel();
         }
 
-        @Override
-        public void setView(View view) {
-            mToast.setView(view);
-        }
-
-        @Override
-        public View getView() {
-            return mToast.getView();
-        }
-
-        @Override
-        public void setDuration(int duration) {
-            mToast.setDuration(duration);
-        }
-
-        @Override
-        public void setGravity(int gravity, int xOffset, int yOffset) {
-            mToast.setGravity(gravity, xOffset, yOffset);
-        }
-
-        @Override
-        public void setText(int resId) {
-            mToast.setText(resId);
-        }
-
-        @Override
-        public void setText(CharSequence s) {
-            mToast.setText(s);
-        }
-
         static class SafeHandler extends Handler {
             private Handler impl;
 
@@ -440,7 +402,7 @@ public final class ToastUtils {
                 try {
                     impl.handleMessage(msg);
                 } catch (Exception e) {
-                    LogUtils.e(e);
+                    Log.e("ToastUtils", e.toString());
                 }
             }
 
@@ -451,9 +413,7 @@ public final class ToastUtils {
         }
     }
 
-    static class ToastWithoutNotification implements IToast {
-
-        private Toast mToast;
+    static class ToastWithoutNotification extends AbsToast {
 
         private WindowManager mWM;
 
@@ -461,20 +421,8 @@ public final class ToastUtils {
 
         private WindowManager.LayoutParams mParams = new WindowManager.LayoutParams();
 
-        private Handler mHandler = new Handler(Looper.myLooper());
-
-        ToastWithoutNotification(@NonNull Toast toast) {
-            mToast = toast;
-
-            mParams.height = WindowManager.LayoutParams.WRAP_CONTENT;
-            mParams.width = WindowManager.LayoutParams.WRAP_CONTENT;
-            mParams.format = PixelFormat.TRANSLUCENT;
-            mParams.windowAnimations = android.R.style.Animation_Toast;
-            mParams.type = WindowManager.LayoutParams.TYPE_TOAST;
-            mParams.setTitle("ToastWithoutNotification");
-            mParams.flags = WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON
-                    | WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE
-                    | WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE;
+        ToastWithoutNotification(Toast toast) {
+            super(toast);
         }
 
         @Override
@@ -484,11 +432,15 @@ public final class ToastUtils {
             Context context = mToast.getView().getContext();
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N_MR1) {
                 mWM = (WindowManager) context.getSystemService(Context.WINDOW_SERVICE);
+                mParams.type = WindowManager.LayoutParams.TYPE_TOAST;
+                mParams.y = mToast.getYOffset();
             } else {
                 Context topActivityOrApp = Utils.getTopActivityOrApp();
                 if (topActivityOrApp instanceof Activity) {
                     mWM = ((Activity) topActivityOrApp).getWindowManager();
                 }
+                mParams.type = WindowManager.LayoutParams.TYPE_APPLICATION_PANEL;
+                mParams.y = mToast.getYOffset() + getNavBarHeight();
             }
 
             final Configuration config = context.getResources().getConfiguration();
@@ -498,6 +450,16 @@ public final class ToastUtils {
             } else {
                 gravity = mToast.getGravity();
             }
+
+            mParams.height = WindowManager.LayoutParams.WRAP_CONTENT;
+            mParams.width = WindowManager.LayoutParams.WRAP_CONTENT;
+            mParams.format = PixelFormat.TRANSLUCENT;
+            mParams.windowAnimations = android.R.style.Animation_Toast;
+
+            mParams.setTitle("ToastWithoutNotification");
+            mParams.flags = WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON
+                    | WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE
+                    | WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE;
             mParams.gravity = gravity;
             if ((gravity & Gravity.HORIZONTAL_GRAVITY_MASK) == Gravity.FILL_HORIZONTAL) {
                 mParams.horizontalWeight = 1.0f;
@@ -506,15 +468,13 @@ public final class ToastUtils {
                 mParams.verticalWeight = 1.0f;
             }
             mParams.x = mToast.getXOffset();
-            mParams.y = mToast.getYOffset();
-
             mParams.packageName = Utils.getApp().getPackageName();
 
             try {
                 mWM.addView(mView, mParams);
             } catch (Exception ignored) { /**/ }
 
-            mHandler.postDelayed(new Runnable() {
+            HANDLER.postDelayed(new Runnable() {
                 @Override
                 public void run() {
                     cancel();
@@ -525,11 +485,32 @@ public final class ToastUtils {
         @Override
         public void cancel() {
             try {
-                mWM.removeView(mView);
-            } catch (IllegalArgumentException ignored) { /**/ }
+                if (mWM != null) {
+                    mWM.removeView(mView);
+                }
+            } catch (Exception ignored) { /**/ }
             mView = null;
-            mHandler = null;
+            mWM = null;
             mToast = null;
+        }
+
+        private int getNavBarHeight() {
+            Resources res = Resources.getSystem();
+            int resourceId = res.getIdentifier("navigation_bar_height", "dimen", "android");
+            if (resourceId != 0) {
+                return res.getDimensionPixelSize(resourceId);
+            } else {
+                return 0;
+            }
+        }
+    }
+
+    static abstract class AbsToast implements IToast {
+
+        Toast mToast;
+
+        AbsToast(Toast toast) {
+            mToast = toast;
         }
 
         @Override